+++ src/matterbridge.bridge.test.ts
@@ -9,7 +9,7 @@ import { jest } from '@jest/globals';
 
 import { AnsiLogger, db, LogLevel, rs, UNDERLINE, UNDERLINEOFF } from 'node-ansi-logger';
 import { Matterbridge } from './matterbridge.js';
-import { waiter } from './utils/utils.js';
+import { waiter } from './utils/export.js';
 import { Environment, StorageService } from '@matter/main';
 import path from 'node:path';
 import os from 'node:os';